Harass: Possible fix for AI appearing at main base (Pleiku etc.).
I think this issue is because a player is marked as harassable and has an AI objective assigned to them.

Then they move back to the main base, but they still have an existing AI objective assigned to them.

The AI objective is never cancelled when they move back to the main base etc.

So AI keep coming to attack them.